+++ CHANGELOG.md
@@ -5,6 +5,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Feature Added: Function `ADDSAssetInventoryAudit` with reporting.
+
 ## [1.9.5] - 2022-11-29
 
 ### Added